requirements.txt
# Core dependencies for OpenAI Agent & MiniMax integration
openai-agents>=0.1.0
litellm>=1.60.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.25.0
pytest-timeout>=2.4.0

# HTTP client for tools
httpx>=0.28.0

# NotebookLM Skill Dependencies
# https://github.com/PleasePrompto/notebooklm-skill
patchright==1.55.2
python-dotenv==1.0.0
